chore(packages): revert consolidate middleware integ tests
This reverts commit 706c2ceff758ff13c491df0c5e5c53d8eec4ce84.
