+++ packages/middleware-apply-body-checksum/jest.config.integ.js
@@ -1,4 +1,4 @@
 module.exports = {
   preset: "ts-jest",
-  testMatch: ["**/*.spec.ts"],
+  testMatch: ["**/*.integ.spec.ts"],
 };
