+++ packages/middleware-endpoint/src/middleware-endpoint.integ.spec.ts
@@ -1,7 +1,7 @@
 import { S3 } from "@aws-sdk/client-s3";
 import { S3Control } from "@aws-sdk/client-s3-control";
 
-import { requireRequestsFrom } from "../../aws-util-test/src";
+import { requireRequestsFrom } from "../../../private/aws-util-test/src";
 
 describe("middleware-endpoint", () => {
   // these are token examples because most endpoint
