--- private/aws-middleware-test/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@aws-sdk/aws-middleware-test",
-  "description": "Integration test suite for AWS middleware",
-  "version": "3.0.0",
-  "scripts": {
-    "build": "exit 0",
-    "build:cjs": "exit 0",
-    "build:es": "exit 0",
-    "build:include:deps": "lerna run --scope $npm_package_name --include-dependencies build",
-    "build:types": "tsc -p tsconfig.types.json",
-    "build:types:downlevel": "downlevel-dts dist-types dist-types/ts3.4",
-    "clean": "rimraf ./dist-* && rimraf *.tsbuildinfo",
-    "test": "jest --coverage --passWithNoTests"
-  },
-  "main": "./dist-cjs/index.js",
-  "types": "./dist-types/index.d.ts",
-  "module": "./dist-es/index.js",
-  "sideEffects": false,
-  "dependencies": {
-    "@aws-sdk/client-accessanalyzer": "*",
-    "@aws-sdk/client-ec2": "*",
-    "@aws-sdk/client-lambda": "*",
-    "@aws-sdk/client-s3": "*",
-    "@aws-sdk/client-s3-control": "*",
-    "@aws-sdk/client-sagemaker": "*",
-    "@aws-sdk/client-sagemaker-runtime": "*",
-    "@aws-sdk/client-xray": "*",
-    "tslib": "^2.5.0"
-  },
-  "devDependencies": {
-    "@tsconfig/node14": "1.0.3",
-    "@types/node": "^12.7.5",
-    "concurrently": "7.0.0",
-    "downlevel-dts": "0.10.1",
-    "typedoc": "0.23.23",
-    "typescript": "~4.9.5"
-  },
-  "overrides": {
-    "typedoc": {
-      "typescript": "~4.9.5"
-    }
-  },
-  "engines": {
-    "node": ">=14.0.0"
-  },
-  "typesVersions": {
-    "<4.0": {
-      "dist-types/*": [
-        "dist-types/ts3.4/*"
-      ]
-    }
-  },
-  "files": [
-    "dist-*/**"
-  ],
-  "author": {
-    "name": "AWS SDK for JavaScript Team",
-    "url": "https://aws.amazon.com/javascript/"
-  },
-  "license": "Apache-2.0",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/aws/aws-sdk-js-v3.git",
-    "directory": "private/aws-middleware-test"
-  }
-}
