--- private/aws-middleware-test/readme.md
@@ -1,10 +1,0 @@
-# @aws-sdk/aws-middleware-test
-
-This is not a runtime or published package.
-
-This is a test spec.
-
-The purpose of this package is to perform integration tests on the AWS middleware.
-
-If tests in this package fail, the author should either fix their changes such that the API contract
-is maintained, or appropriately announce and safely deprecate the interfaces affected by incoming changes.
